+++ src/py/loaders/__init__.py
@@ -3,3 +3,4 @@ from src.py.loaders.fileloaders import loadBinaryImage,loadIntensityImage,loadDV
     loadDVMultiChannelImage,loadTifMultiChannelImage
 from src.py.loaders.cellloader import loadCells
 from src.py.loaders.modelloader import loadModel
+from src.py.loaders.maskloader import loadMaskFile,loadMaskFileFromBinaryImage